+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.2.0] - 2025-06-15
+
 ### Added
 
 - Added binary mode to PipeProcess via new binary_content parameter